# server/openapi.yaml
openapi: 3.0.3
info:
  title: OpenBin API
  version: 1.0.0
  description: |
    REST API for OpenBin — a multi-user web app for organizing physical storage bins with QR codes.

    **Authentication**: Most endpoints require authentication via one of:
    - **httpOnly cookies** (primary) — `openbin-access` (15 min) + `openbin-refresh` (7 days), set automatically on login/register. Browsers send these with every request.
    - **API key** created via `POST /api-keys` (long-lived, per-user, format: `sk_openbin_...`) — `Authorization: Bearer <key>`. A key grants access to all locations the user belongs to.

    **Response envelopes**: All list endpoints return `{ results: T[], count: number }`.
    Error responses use `{ error: "CODE", message: "Human readable" }`.

    **Rate limiting**: All API endpoints share a global limit of 200 requests/minute.
servers:
  - url: /api
    description: API base path (proxied via nginx)
security:
  - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: API key (sk_openbin_...) via Authorization header. Cookies are the primary auth method for browsers.
  schemas:
...
    Location:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        created_by:
          type: string
          format: uuid
          nullable: true
        invite_code:
          type: string
        activity_retention_days:
          type: integer
          minimum: 7
          maximum: 365
          default: 90
          description: Days to keep activity log entries before auto-pruning
        trash_retention_days:
          type: integer
          minimum: 7
          maximum: 365
          default: 30
          description: Days to keep soft-deleted bins before auto-purging
        app_name:
          type: string
          default: OpenBin
          description: Custom display name for this location's app instance
        term_bin:
          type: string
          default: bin
          description: Custom terminology for 'bin'
        term_location:
          type: string
          default: location
          description: Custom terminology for 'location'
        term_area:
          type: string
          default: area
          description: Custom terminology for 'area'
        default_join_role:
          type: string
          enum: [member, viewer]
          default: member
          description: Default role assigned to new members joining via invite code
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    LocationWithMeta:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        created_by:
          type: string
          format: uuid
          nullable: true
        invite_code:
          type: string
        activity_retention_days:
          type: integer
          minimum: 7
          maximum: 365
          default: 90
        trash_retention_days:
          type: integer
          minimum: 7
          maximum: 365
          default: 30
        app_name:
          type: string
          default: OpenBin
        term_bin:
          type: string
          default: bin
        term_location:
          type: string
          default: location
        term_area:
          type: string
          default: area
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        member_count:
          type: integer
        area_count:
          type: integer
        bin_count:
          type: integer
          description: Number of non-deleted bins in this location
        role:
          type: string
          enum: [admin, member, viewer]
        default_join_role:
          type: string
          enum: [member, viewer]
          default: member
          description: Default role assigned to new members joining via invite code
...
    CreateLocationRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          maxLength: 255
...
    BinVisibility:
      type: string
      enum: [location, private]
      default: location
      description: >
        Visibility level for a bin.
        `location` — visible to all location members (default).
        `private` — visible only to the creator.
    CreateBinRequest:
      type: object
      required: [name, locationId]
      properties:
        name:
          type: string
          maxLength: 255
        locationId:
          type: string
          format: uuid
        items:
          type: array
          items:
            type: string
          maxItems: 500
        notes:
          type: string
          maxLength: 10000
        tags:
          type: array
          items:
            type: string
          maxItems: 50
        areaId:
          type: string
          format: uuid
        icon:
          type: string
        color:
          type: string
        cardStyle:
          type: string
          maxLength: 500
          description: JSON-encoded card style configuration
        visibility:
          $ref: '#/components/schemas/BinVisibility'
        customFields:
          type: object
          additionalProperties:
            type: string
          description: "Custom field values as { fieldId: value } map"
...
  responses:
    Unauthorized:
      description: Missing or invalid authentication
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: UNAUTHORIZED
            message: Not authenticated
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: FORBIDDEN
            message: Not a member of this location
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: NOT_FOUND
            message: Resource not found
    ValidationError:
      description: Invalid request parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: VALIDATION_ERROR
            message: Name is required
...
paths:
...
  /locations:
    get:
      tags: [Locations]
      summary: List user's locations
      description: Returns locations the current user is a member of, with member count and role.
      responses:
        '200':
          description: List of locations
          content:
            application/json:
              schema:
                type: object
                properties:
                  results:
                    type: array
                    items:
                      $ref: '#/components/schemas/LocationWithMeta'
                  count:
                    type: integer
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [Locations]
      summary: Create a new location
      description: Creates a location and adds the current user as owner.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateLocationRequest'
      responses:
        '201':
          description: Location created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Location'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationError'
...
  /locations/{id}/members/{userId}:
    delete:
      tags: [Locations]
      summary: Remove a member from location
      description: Admins can remove anyone. Members can only remove themselves (leave). Last admin cannot leave.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Member removed
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Member removed
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /locations/{id}/members/{userId}/role:
    put:
      tags: [Locations]
      summary: Change a member's role
      description: Admin only. Change a member's role to admin, member, or viewer. Cannot demote the last admin.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [role]
              properties:
                role:
                  type: string
                  enum: [admin, member, viewer]
      responses:
        '200':
          description: Role updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Role updated to admin
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/ValidationError'
...
  /bins:
    post:
      tags: [Bins]
      summary: Create a bin
      description: |
        Creates a new bin with an auto-generated 6-character alphanumeric ID.
        The ID uses a charset excluding ambiguous characters (0/O, 1/l/I).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBinRequest'
      responses:
        '201':
          description: Bin created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Bin'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/ValidationError'
    get:
      tags: [Bins]
      summary: List bins for a location
      description: |
        Returns non-deleted bins. Supports server-side search, filtering, sorting,
        and optional offset-based pagination.

        When `limit` is provided, responses include a `count` field with the total
        number of matching rows (for building pagination UI). When `limit` is absent,
        all matching bins are returned and `count` equals `results.length` (backward
        compatible).
      parameters:
        - name: location_id
          in: query
          required: true
          schema:
            type: string
            format: uuid
        - name: q
          in: query
          schema:
            type: string
          description: Search across bin name, items, notes, tags, ID, and area name
        - name: tag
          in: query
          schema:
            type: string
          description: Filter by exact tag match (use `tags` for multi-tag)
        - name: tags
          in: query
          schema:
            type: string
          description: Comma-separated tag names. Combined with `tag_mode`.
        - name: tag_mode
          in: query
          schema:
            type: string
            enum: [any, all]
            default: any
          description: When filtering by `tags`, require any or all tags to match
        - name: area_id
          in: query
          schema:
            type: string
          description: "Filter by single area. Use '__unassigned__' for bins with no area. See also `areas`."
        - name: areas
          in: query
          schema:
            type: string
          description: "Comma-separated area IDs. Supports '__unassigned__'. Takes precedence over `area_id`."
        - name: colors
          in: query
          schema:
            type: string
          description: Comma-separated color keys to filter by
        - name: has_items
          in: query
          schema:
            type: string
            enum: ['true']
          description: Only return bins that have at least one item
        - name: has_notes
          in: query
          schema:
            type: string
            enum: ['true']
          description: Only return bins with non-empty notes
        - name: needs_organizing
          in: query
          schema:
            type: string
            enum: ['true']
          description: Filter bins missing tags, area, and items
        - name: sort
          in: query
          schema:
            type: string
            enum: [name, created_at, updated_at, area]
            default: updated_at
        - name: sort_dir
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
          description: Page size. When provided, enables pagination and `count` returns total matching rows.
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
          description: Number of rows to skip (used with `limit`)
      responses:
        '200':
          description: List of bins
          content:
            application/json:
              schema:
                type: object
                properties:
                  results:
                    type: array
                    items:
                      $ref: '#/components/schemas/Bin'
                  count:
                    type: integer
                    description: Total matching rows when paginated, or results.length otherwise
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /bins/{id}:
    get:
      tags: [Bins]
      summary: Get a single bin
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            description: 6-character alphanumeric bin short code
      responses:
        '200':
          description: Bin details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Bin'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Bins]
      summary: Update a bin
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            description: 6-character alphanumeric bin short code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateBinRequest'
      responses:
        '200':
          description: Bin updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Bin'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
    delete:
      tags: [Bins]
      summary: Soft-delete a bin
      description: Sets deleted_at timestamp. Bin can be restored from trash.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            description: 6-character alphanumeric bin short code
      responses:
        '200':
          description: Bin soft-deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Bin deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...